Delete mock-only files
Removed:
- lib/finance-governance/mock-data.ts (mock data generators)
- lib/finance-governance/server-memory-store.ts (mock store, unused)
- tests/harness/connectors/mock-stripe.ts (test mock)
- tests/harness/connectors/mock-github.ts (test mock)
- tests/harness/connectors/mock-vercel.ts (test mock)

Created:
- lib/finance-governance/types.ts (extracted type definitions)

Updated imports in repository.ts to use types.ts instead of mock-data.ts.
